+++ lib/web_utils.dart
@@ -7,3 +7,4 @@ export 'package:js_interop_utils/js_interop_utils.dart';
 export 'package:web/web.dart';
 
 export 'src/web_utils_extensions.dart';
+export 'src/web_utils_helpers.dart';